Make BlockwiseLint work for liturgy too, and factor out SpellCheck.
